Tried to get the entire test suite running, but this is impossible due to GAE constraints concerning a limit of 5 entity groups per tx.
Former-commit-id: 477e629ccd3b7004ad2da3dcc39c91938a8eded9

+++ org.cumulus4j/org.cumulus4j.store.test/src/main/java/org/cumulus4j/store/test/framework/AbstractJDOTransactionalTestClearingDatabase.java
@@ -8,8 +8,14 @@ public class AbstractJDOTransactionalTestClearingDatabase extends AbstractJDOTra
 {
 	private static final Logger logger = LoggerFactory.getLogger(AbstractJDOTransactionalTestClearingDatabase.class);
 
+	public static boolean clearDatabaseEnabled = true;
+
 	@BeforeClass
 	public static void clearDatabase() throws Exception {
+		if (!clearDatabaseEnabled) {
+			logger.info("clearDatabase: Clearing database is DISABLED! Skipping.");
+			return;
+		}
 		logger.info("clearDatabase: Clearing database (dropping all tables).");
 		CleanupUtil.dropAllTables();
 	}
